feat: add javadoc for module Node
