+++ common/src/main/java/com/github/jon7even/telegram/menu/MainMenu.java
@@ -14,26 +14,32 @@ public enum MainMenu {
      * Регистрация и инициализация пользователя
      */
     START("/start"),
+
     /**
      * Выводит помощь по доступным командам
      */
     HELP("/help"),
+
     /**
      * Инициализирует работу с товарами
      */
     ITEMS("/items"),
+
     /**
      * Инициализирует работу с задачами
      */
     TASKS("/tasks"),
+
     /**
      * Инициализирует работу с конкурентами
      */
     COMPETITORS("/competitors"),
+
     /**
      * Инициализирует работу с подарками
      */
     GIFTS("/gifts");
+
     private final String command;
 
     @Override
