+++ node/src/main/java/com/github/jon7even/NodeApp.java
@@ -3,6 +3,12 @@ package com.github.jon7even;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
+/**
+ * Основной класс запуска сервиса "Узел", управляется фреймворком SpringBoot
+ *
+ * @author Jon7even
+ * @version 1.0
+ */
 @SpringBootApplication
 public class NodeApp {
     public static void main(String[] args) {
