+++ node/src/main/java/com/github/jon7even/configuration/RabbitConfiguration.java
@@ -5,6 +5,12 @@ import org.springframework.amqp.support.converter.MessageConverter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
+/**
+ * Конфигурация для работы с брокером очередей RabbitMq
+ *
+ * @author Jon7even
+ * @version 1.0
+ */
 @Configuration
 public class RabbitConfiguration {
     @Bean
