+++ node/src/main/java/com/github/jon7even/dto/company/CompanyBuildingDto.java
@@ -5,6 +5,12 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+/**
+ * Класс DTO для создания новой компании
+ *
+ * @author Jon7even
+ * @version 1.0
+ */
 @Data
 @Builder
 @AllArgsConstructor
