+++ node/src/main/java/com/github/jon7even/dto/gift/GiftShortDto.java
@@ -6,6 +6,12 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+/**
+ * Класс DTO для краткого представления подарка
+ *
+ * @author Jon7even
+ * @version 1.0
+ */
 @Data
 @Builder
 @AllArgsConstructor
