+++ node/src/main/java/com/github/jon7even/entity/competitor/CompetitorEntity.java
@@ -6,6 +6,12 @@ import lombok.*;
 
 import java.time.LocalDateTime;
 
+/**
+ * Класс описывающий сущность конкурента
+ *
+ * @author Jon7even
+ * @version 1.0
+ */
 @Entity
 @Builder
 @Setter
