+++ node/src/main/java/com/github/jon7even/entity/event/EventEntity.java
@@ -6,6 +6,12 @@ import lombok.*;
 
 import java.time.LocalDateTime;
 
+/**
+ * Класс описывающий сущность события
+ *
+ * @author Jon7even
+ * @version 1.0
+ */
 @Entity
 @Builder
 @Setter
