+++ node/src/main/java/com/github/jon7even/mapper/CompanyMapper.java
@@ -1,15 +1,21 @@
 package com.github.jon7even.mapper;
 
 import com.github.jon7even.dto.company.CompanyBuildingDto;
-import com.github.jon7even.entity.user.UserEntity;
 import com.github.jon7even.entity.company.CompanyEntity;
 import com.github.jon7even.entity.gift.GiftEntity;
+import com.github.jon7even.entity.user.UserEntity;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.factory.Mappers;
 
 import java.time.LocalDateTime;
 
+/**
+ * Интерфейс для маппинга DTO и сущностей компаний
+ *
+ * @author Jon7even
+ * @version 1.0
+ */
 @Mapper(componentModel = "spring")
 public interface CompanyMapper {
     CompanyMapper INSTANCE = Mappers.getMapper(CompanyMapper.class);
