+++ node/src/main/java/com/github/jon7even/mapper/EventMapper.java
@@ -2,12 +2,18 @@ package com.github.jon7even.mapper;
 
 import com.github.jon7even.dto.event.EventBuildingDto;
 import com.github.jon7even.dto.event.EventShortDto;
-import com.github.jon7even.entity.user.UserEntity;
 import com.github.jon7even.entity.event.EventEntity;
+import com.github.jon7even.entity.user.UserEntity;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.factory.Mappers;
 
+/**
+ * Интерфейс для маппинга DTO и сущностей событий
+ *
+ * @author Jon7even
+ * @version 1.0
+ */
 @Mapper(componentModel = "spring")
 public interface EventMapper {
     EventMapper INSTANCE = Mappers.getMapper(EventMapper.class);
