+++ node/src/main/java/com/github/jon7even/mapper/GiftMapper.java
@@ -2,12 +2,18 @@ package com.github.jon7even.mapper;
 
 import com.github.jon7even.dto.gift.GiftBuildingDto;
 import com.github.jon7even.dto.gift.GiftShortDto;
-import com.github.jon7even.entity.user.UserEntity;
 import com.github.jon7even.entity.gift.GiftEntity;
+import com.github.jon7even.entity.user.UserEntity;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.factory.Mappers;
 
+/**
+ * Интерфейс для маппинга DTO и сущностей подарков
+ *
+ * @author Jon7even
+ * @version 1.0
+ */
 @Mapper(componentModel = "spring")
 public interface GiftMapper {
     GiftMapper INSTANCE = Mappers.getMapper(GiftMapper.class);
