+++ node/src/main/java/com/github/jon7even/service/in/consumer/ConsumerService.java
@@ -2,14 +2,36 @@ package com.github.jon7even.service.in.consumer;
 
 import org.telegram.telegrambots.meta.api.objects.Update;
 
+/**
+ * Интерфейс сервиса слушателя очередей RabbitMq входящих сообщений от Telegram для их дальнейшей обработки
+ *
+ * @author Jon7even
+ * @version 1.0
+ * @apiNote Основная точка поступления данных в сервис "Узел"
+ */
 public interface ConsumerService {
+    /**
+     * Метод для обработки входящих текстовых сообщений
+     */
     void consumeTextMessageUpdates(Update update);
 
+    /**
+     * Метод для обработки процесса нажатия на клавиатуру
+     */
     void consumeCallbackQueryUpdates(Update update);
 
+    /**
+     * Метод для обработки входящих документов
+     */
     void consumeDocMessageUpdates(Update update);
 
+    /**
+     * Метод для обработки входящих аудио сообщений
+     */
     void consumeAudioMessageUpdates(Update update);
 
+    /**
+     * Метод для обработки входящих фото
+     */
     void consumePhotoMessageUpdates(Update update);
 }
