+++ node/src/main/java/com/github/jon7even/telegram/BotState.java
@@ -1,11 +1,23 @@
 package com.github.jon7even.telegram;
 
+/**
+ * Перечисление текущего состояния бота для конкретного пользователя
+ *
+ * @author Jon7even
+ * @version 1.0
+ */
 public enum BotState {
+    /**
+     * Статусы главного меню
+     */
     MAIN_START,
     MAIN_HELP,
     MAIN_ITEMS,
     MAIN_GIFTS,
 
+    /**
+     * Статусы добавления новой компании
+     */
     COMPANY_NAME,
     COMPANY_SUM,
     COMPANY_TYPE_GIFT,
