--- node/src/main/java/com/github/jon7even/utils/Emoji.java
@@ -1,16 +1,0 @@
-package com.github.jon7even.utils;
-
-import com.vdurmont.emoji.EmojiParser;
-import lombok.AllArgsConstructor;
-
-@AllArgsConstructor
-public enum Emoji {
-    SMAIL(EmojiParser.parseToUnicode(":blush:"));
-
-    private final String emojiCode;
-
-    @Override
-    public String toString() {
-        return emojiCode;
-    }
-}
